Correct release date for 3.2.1
Previously, the release date for version 3.2.1 was listed as 2018-05-21 in
the release notes, but the actual release date was 2018-05-28.  2018-05-21
was the date of the previous release, so it looks like a simple copy/paste
error.

This commit updates the listed release date to match the actual date of
the release.