Use exit codes between 0 and 255 (inclusive)
Shells express exit codes in the range of 0-255. Anything outside of that range is brought into that range by modding it 256.

For example:
-1 becomes 255
-2 becomes 254
-3 becomes 253

This need to convert exit codes in shell scripts to use dependency check is unexpected and confusing.

For that reason, returning exit codes outside of 0-255 is unusual and generally discouraged. Therefore, use exit codes in the 0-255 range.

With this change, exit code 1 was used for both help and CVSS score
failure. Therefore, the CVSS score failure exit code was changed to 15.

See: https://tldp.org/LDP/abs/html/exit-status.html